module5-task1 модуляция начальный варинат
